[Refactor] 지도 커스텀 훅 추출
- 기능별 useEffect 코드 → 커스텀 훅으로 분리
